+++ apps/server/package.json
@@ -30,7 +30,6 @@
 		"@snowcrystals/icicle": "^2.0.4",
 		"axios": "^1.5.0",
 		"body-parser": "^1.20.2",
-		"busboy": "^1.6.0",
 		"colorette": "^2.0.20",
 		"cookie-parser": "^1.4.6",
 		"cors": "2.8.5",
@@ -44,7 +43,6 @@
 		"lodash": "^4.17.21",
 		"mime-types": "^2.1.35",
 		"ms": "^2.1.3",
-		"multer": "^1.4.5-lts.1",
 		"node-2fa": "^2.0.3",
 		"node-os-utils": "^1.3.7",
 		"pidusage": "^3.0.2",
@@ -56,7 +54,6 @@
 		"zip-lib": "^0.7.3"
 	},
 	"devDependencies": {
-		"@types/busboy": "^1.5.0",
 		"@types/cookie-parser": "^1.4.4",
 		"@types/cors": "2.8.14",
 		"@types/cron": "^2.0.1",
@@ -67,7 +64,6 @@
 		"@types/lodash": "^4.14.198",
 		"@types/mime-types": "^2.1.1",
 		"@types/ms": "^0.7.31",
-		"@types/multer": "^1.4.7",
 		"@types/node": "^18.17.15",
 		"@types/node-os-utils": "^1.3.1",
 		"@types/pidusage": "^2.0.2",
